Change `-j` argument to be more generic
Make allows for a nice trick when specifying how how many jobs to use. If you pass `-j`without a number it allows make to use as many jobs as it can. But if you pass the `-l` flag it will prevent make from spawning more jobs if the CPU load is at or above the number you specify. (Just in case, CPU load isn't 0-100 like on Windows, it is 0-N where N is the count of hardware threads you have. So on machine with 10 threads a load of 1.1 would be 11%) This makes it so make won't completely overwhelm your system and allows you to do other things, since these other things put load on the CPU.

The CPU limit I provide with this change comes from the `nproc` command which gets the number of threads the system has. This number is provided to make by using the backtick syntax which replaces the part of the commamd surrounded by backytcks (inclusive) with the output of the executed command.

I'm sorry if this is over explaining, I just found your content and I'm not yet familiar with how much you know about these tools.